Refactor secret helpers and credential envelope validation
