+++ internal/vaults/handler.go
@@ -156,11 +156,11 @@ func SealCredentialSecret(ctx context.Context, secretSvc *secrets.Service, crede
 	return nil
 }
 
-// OpenCredentialSecret decrypts a credential's envelope back into SecretPayload
+// openCredentialSecret decrypts a credential's envelope back into SecretPayload
 // so callers can read the secret for merge or runtime use. A missing envelope
 // returns ErrMissingSecretEnvelope. A present envelope that fails to open
 // fails closed. Decrypted plaintext must not be persisted or logged.
-func OpenCredentialSecret(ctx context.Context, secretSvc *secrets.Service, credential *db.VaultCredential) error {
+func openCredentialSecret(ctx context.Context, secretSvc *secrets.Service, credential *db.VaultCredential) error {
 	if credential.SecretEnvelope == nil {
 		return ErrMissingSecretEnvelope
 	}
@@ -648,7 +648,7 @@ func (h *Handler) updateCredentialRoute(w http.ResponseWriter, r *http.Request) 
 		// A missing envelope can still be repaired when the body carries a
 		// complete replacement secret; otherwise ask the client to resubmit.
 		if current.SecretEnvelope != nil {
-			if err := OpenCredentialSecret(r.Context(), h.secretSvc, &current); err != nil {
+			if err := openCredentialSecret(r.Context(), h.secretSvc, &current); err != nil {
 				writeSecretOpenError(w, r, h.logger, "update", err)
 				return
 			}
@@ -788,7 +788,7 @@ func (h *Handler) validateCredentialRoute(w http.ResponseWriter, r *http.Request
 	}
 	// Decrypt transiently to inspect whether a refresh token is present. The
 	// plaintext is not persisted or logged.
-	if err := OpenCredentialSecret(r.Context(), h.secretSvc, &credential); err != nil {
+	if err := openCredentialSecret(r.Context(), h.secretSvc, &credential); err != nil {
 		writeSecretOpenError(w, r, h.logger, "validate", err)
 		return
 	}
